// frontend/src/types/auth.ts
export interface User {
  id: string
  email: string
  name: string | null
}

export interface AuthResponse {
  message?: string
  token: string
  user: User
}

export interface AuthContextType {
  user: User | null
  token: string | null
  isAuthenticated: boolean
  isLoading: boolean
  login: (email: string, password: string) => Promise<void>
  register: (name: string, email: string, password: string) => Promise<void>
  logout: () => void
}

export interface AuthMeResponse {
  user: User
}